build: use xxhsum if available. fall back on sha1sum if not
xxhsum is much faster than sha1sum. it's from the xxhash project:
https://github.com/Cyan4973/xxHash

it's a non-cryptographic hashing algorithm, running at RAM speed limits. the
priority in untitled is speed. collisions are extremely unlikely

I'm going through untitled to see what else I can optimize. right now, the
biggest bottleneck is disk I/O. what I can do to avoid use of disk I/O could
be to rely on a git repository, if available on a given site, and use that
git repository to keep track of changes to a site. that way, it'd be very fast
to just keep a site up to date. (and fall back to current behaviour if that
site isn't in a git repository)