treat tabs as 8 spaces
when showing code/config samples, tab-based indentation shows 4
spaces when rendered as html

this is contrary to how files are typically displayed, on terminals
in posix systems. for example, if outputting a .c file with tab
indentation in it, using the "cat" program, tabs are represented as
being the same visual width as 8 spaces

pandoc converts tabs to spaces, but treats them as 4 spaces. this
change makes it treat them as 8 spaces